move booru code to service
